refac: detailView rename
